Upgrade to pdfbox 3.0.1
It means that FontVerter font conversion has to be disabled. It is not
working anymore as it only works with fontbox v2 and not v3.

Also bump:

maven-compiler-plugin to 3.12.1
maven-jar-plugin to 3.3.0
maven-resources-plugin to 3.3.1
maven-surefire-plugin to 3.2.5
maven-assembly-plugin to 3.6.0
maven-source-plugin to 3.3.0
maven-release-plugin to 3.0.1
nexus-staging-maven-plugin to 1.6.13
maven-javadoc-plugin to 3.6.3
maven-gpg-plugin to 3.1.0

commons-io to 2.15.1
slf4j-api to 2.0.12
junit to 5.10.2
jsoup to 1.17.2
hamcrest to 2.2
commons-codec to 1.16.1
slf4j to 2.0.12